+++ quix-frontend/client/src/services/subscription.ts
@@ -1,0 +1,20 @@
+import cookie from 'cookie';
+import {ClientConfigHelper} from '@wix/quix-shared';
+import {Store, sessionId} from '../lib/store';
+import {App} from '../lib/app';
+
+export const subscribeToStateChanges = (app: App<ClientConfigHelper>, store: Store) => {
+  const ws = new WebSocket(`ws://${location.host}/subscription`);
+    
+  const {authCookieName} = app.getConfig().getAuth();
+  const cookies = cookie.parse(document.cookie);
+  const token = cookies[authCookieName];
+
+  ws.onopen = () => {
+    ws.onmessage = (message: MessageEvent) => {
+      store.dispatch(JSON.parse(message.data).data);
+    };
+
+    ws.send(JSON.stringify({ event: 'subscribe', data: {token, sessionId} }));
+  };
+}
