+++ quix-frontend/client/src/state-components/base/base.ts
@@ -1,13 +1,11 @@
 import template from './base.html';
 
-import {Store, sessionId} from '../../lib/store';
+import {ClientConfigHelper} from '@wix/quix-shared';
+import {Store} from '../../lib/store';
 import {App} from '../../lib/app';
 import {IStateComponentConfig} from '../../lib/app/services/plugin-builder';
 import {setSearchText, setSearchPage} from '../../store/app/app-actions';
-import {openSearchResults, closePopup, hasQueuedNotes} from '../../services';
-import { ClientConfigHelper } from '@wix/quix-shared';
-import cookie from 'cookie';
-
+import {openSearchResults, closePopup, hasQueuedNotes, subscribeToStateChanges} from '../../services';
 
 export default (app: App<ClientConfigHelper>, store: Store) => ({
   name: '',
@@ -20,18 +18,7 @@ export default (app: App<ClientConfigHelper>, store: Store) => ({
   scope: {},
   controller: (scope, params, {syncUrl}) => {
     syncUrl();
-
-    const ws = new WebSocket(`ws://${location.host}/subscription`);
-    
-    const {authCookieName} = app.getConfig().getAuth();
-    const cookies = cookie.parse(document.cookie);
-    const token = cookies[authCookieName];
-    ws.onopen = () => {
-      ws.onmessage = (message: MessageEvent) => {
-        store.dispatch(JSON.parse(message.data).data)
-      };
-      ws.send(JSON.stringify({ event: 'subscribe', data: {token, sessionId} }));
-    };
+    subscribeToStateChanges(app, store);
 
     store.subscribe('app.searchText', text => text ? openSearchResults(scope) : closePopup(), scope);
